requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# Facebook Ads
facebook-business==18.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Logging
colorlog==6.8.0

# Testing (optional)
pytest==7.4.3
